Time: 11 ms (100.00%), Space: 5.9 MB (100.00%) - LeetHub

+++ 2240-number-of-ways-to-buy-pens-and-pencils/2240-number-of-ways-to-buy-pens-and-pencils.cpp
@@ -10,7 +10,7 @@ public:
         
         for(int i=0;i<=lim;i++)
         {
-            res = res + ((total - (long long)(i*fir))/sec + 1);
+            res = res + ((total - (i*fir))/sec + 1);
         }
         
         return res;
